+++ quest/include/operations.h
@@ -1,15 +1,19 @@
 /** @file
- * API signatures for effecting operators (such as gates and unitaries) 
- * upon Quregs which are instantiated as either statevectors or 
- * density matrices. This excludes decoherence channels which are
- * instead exposed in decoherence.h
+ * API signatures for effecting mostly physical and/or trace
+ * preserving operators, such as unitaries, gates and 
+ * measurements, upon Quregs which are instantiated as both 
+ * statevectors or density matrices. This excludes Trotterised
+ * gadgets and evolutions (exposed instead in trotterisation.h),
+ * functions to pre- or post-multiply operators upon density
+ * matrices (multiplication.h) and decoherence channels
+ * (decoherence.h).
  * 
  * @author Tyson Jones
  * @author Diogo Pratas Maia (non-unitary Pauli gadget)
  * 
  * @defgroup operations Operations
  * @ingroup api
- * @brief Functions for effecting operators upon Quregs.
+ * @brief Functions for effecting standard operators upon Quregs.
  * @{
  */
 
@@ -44,108 +48,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/** Multiplies a general one-qubit dense @p matrix upon the specified @p target 
- * qubit of @p qureg.
- *  
- * @formulae
- * Let @f$ \hat{M} = @f$ @p matrix and @f$ t = @f$ @p target, and notate 
- * @f$\hat{M}_t@f$ as per applyCompMatr1(). Unlike applyCompMatr1() however,
- * this function only ever left-multiplies @p matrix upon @p qureg, regardless
- * of whether it is a statevector or density matrix.
- * 
- * Explicitly,
- * - When @p qureg is a statevector @f$ \svpsi @f$, this function effects
- *   @f[ 
-        \svpsi \rightarrow \hat{M}_t \, \svpsi.
- *   @f]
- * - When @p qureg is a density matrix @f$\dmrho@f$, this function effects
- *   @f[ 
-        \dmrho \rightarrow \hat{M}_t \, \dmrho.
- *   @f]
- *
- * There are no additional constraints like unitarity.
- *
- * @myexample
- * ```
-    Qureg qureg = createDensityQureg(5);
-
-    CompMatr1 matrix = getInlineCompMatr1({
-        {0.1, 0.2},
-        {0.3i, 0.4i}
-    });
-
-    multiplyCompMatr1(qureg, 2, matrix); 
- * ```
- *
- * @param[in,out] qureg  the state to modify.
- * @param[in]     target the index of the target qubit.
- * @param[in]     matrix the Z-basis matrix to multiply.
- * @throws @validationerror
- * - if @p qureg or @p matrix are uninitialised.
- * - if @p target is an invalid qubit index.
- * @see
- * - getCompMatr1()
- * - getInlineCompMatr1()
- * - applyCompMatr1()
- * - postMultiplyCompMatr1()
- * - applyQubitProjector()
- * - multiplyCompMatr()
- * @author Tyson Jones
- */
-void multiplyCompMatr1(Qureg qureg, int target, CompMatr1 matrix);
-
-
-/** @notyettested
- * 
- * Multiplies a general one-qubit dense @p matrix upon the specified @p target 
- * qubit of the density matrix @p qureg, from the right-hand side.
- *  
- * @formulae
- * Let @f$ \dmrho = @f$ @p qureg, @f$ \hat{M} = @f$ @p matrix and @f$ t = @f$ @p target, 
- * and notate @f$\hat{M}_t@f$ as per applyCompMatr1(). Unlike applyCompMatr1() however,
- * this function only ever right-multiplies @p matrix upon @p qureg.
- * 
- * Explicitly
- *   @f[ 
-        \dmrho \rightarrow \dmrho \, \hat{M}_t
- *   @f]
- * where @f$ \hat{M} @f$ is not conjugated nor transposed, and there are no additional 
- * constraints like unitarity.
- * 
- * In general, this function will break the normalisation of @p qureg and result in a
- * non-physical state, and is useful for preparing sub-expressions of formulae like
- * the Linbladian.
- *
- * @myexample
- * ```
-    Qureg qureg = createDensityQureg(5);
-
-    CompMatr1 matrix = getInlineCompMatr1({
-        {0.1, 0.2},
-        {0.3i, 0.4i}
-    });
-
-    postMultiplyCompMatr1(qureg, 2, matrix); 
- * ```
- *
- * @param[in,out] qureg  the state to modify.
- * @param[in]     target the index of the target qubit.
- * @param[in]     matrix the Z-basis matrix to post-multiply.
- * @throws @validationerror
- * - if @p qureg or @p matrix are uninitialised.
- * - if @p qureg is not a density matrix.
- * - if @p target is an invalid qubit index.
- * @see
- * - getCompMatr1()
- * - getInlineCompMatr1()
- * - applyCompMatr1()
- * - multiplyCompMatr1()
- * - multiplyCompMatr()
- * @author Tyson Jones
- */
-void postMultiplyCompMatr1(Qureg qureg, int target, CompMatr1 matrix);
 
 
 /** Applies a general one-qubit dense unitary @p matrix to the specified @p target 
@@ -392,21 +294,6 @@ extern "C" {
 #endif
 
 
-/// @notyetdoced
-/// @see
-/// - applyCompMatr2()
-/// - multiplyCompMatr1()
-void multiplyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matr);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-/// @see
-/// - postMultiplyCompMatr1
-void postMultiplyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matrix);
-
-
 /** @notyetdoced
  * 
  * Applies a general two-qubit dense unitary @p matrix to qubits @p target1 and
@@ -435,6 +322,8 @@ digraph {
  *
  * @see
  * - applyCompMatr1()
+ * - multiplyCompMatr2()
+ * - postMultiplyCompMatr2()
  */
 void applyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matrix);
 
@@ -611,23 +500,6 @@ extern "C" {
 
 /** @notyetdoced
  * 
- * @see
- * - applyCompMatr()
- * - multiplyCompMatr1()
- */
-void multiplyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matrix);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-/// @see
-/// - postMultiplyCompMatr1
-void postMultiplyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matrix);
-
-
-/** @notyetdoced
- * 
  * @formulae
  * 
  * Let @f$ M = @f$ @p matrix.
@@ -643,6 +515,8 @@ void postMultiplyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr ma
  *
  * @see
  * - applyCompMatr1()
+ * - multiplyCompMatr()
+ * - postMultiplyCompMatr()
  */
 void applyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matr);
 
@@ -671,22 +545,6 @@ void applyMultiStateControlledCompMatr(Qureg qureg, int* controls, int* states, 
 #endif
 
 #ifdef __cplusplus
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see multiplyCompMatr()
-void multiplyCompMatr(Qureg qureg, std::vector<int> targets, CompMatr matr);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see postMultiplyCompMatr()
-void postMultiplyCompMatr(Qureg qureg, std::vector<int> targets, CompMatr matr);
 
 
 /// @notyettested
@@ -739,19 +597,12 @@ extern "C" {
 #endif
 
 
-/// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplyDiagMatr1(Qureg qureg, int target, DiagMatr1 matr);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-void postMultiplyDiagMatr1(Qureg qureg, int target, DiagMatr1 matrix);
-
-
-/// @notyetdoced
-/// @see applyCompMatr1()
+/** @notyetdoced
+ * @see 
+ * - applyCompMatr1()
+ * - multiplyCompMatr2()
+ * - postMultiplyCompMatr2()
+ */
 void applyDiagMatr1(Qureg qureg, int target, DiagMatr1 matr);
 
 
@@ -810,17 +661,6 @@ void applyMultiStateControlledDiagMatr1(Qureg qureg, std::vector<int> controls, 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplyDiagMatr2(Qureg qureg, int target1, int target2, DiagMatr2 matr);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-void postMultiplyDiagMatr2(Qureg qureg, int target1, int target2, DiagMatr2 matrix);
 
 
 /// @notyetdoced
@@ -886,17 +726,6 @@ extern "C" {
 
 
 /// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-void postMultiplyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix);
-
-
-/// @notyetdoced
 /// @see applyCompMatr1()
 void applyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix);
 
@@ -914,19 +743,6 @@ void applyMultiControlledDiagMatr(Qureg qureg, int* controls, int numControls, i
 /// @notyetdoced
 /// @see applyMultiStateControlledCompMatr1()
 void applyMultiStateControlledDiagMatr(Qureg qureg, int* controls, int* states, int numControls, int* targets, int numTargets, DiagMatr matrix);
-
-
-/// @notyetdoced
-/// @see
-/// - multiplyCompMatr1()
-/// - applyDiagMatrPower()
-void multiplyDiagMatrPower(Qureg qureg, int* targets, int numTargets, DiagMatr matrix, qcomp exponent);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-void postMultiplyDiagMatrPower(Qureg qureg, int* targets, int numTargets, DiagMatr matrix, qcomp exponent);
 
 
 /** @notyetdoced
@@ -971,22 +787,6 @@ void applyMultiStateControlledDiagMatrPower(Qureg qureg, int* controls, int* sta
 /// @notyetvalidated
 /// @notyetdoced
 /// @cppvectoroverload
-/// @see multiplyDiagMatr()
-void multiplyDiagMatr(Qureg qureg, std::vector<int> targets, DiagMatr matrix);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see postMultiplyDiagMatr()
-void postMultiplyDiagMatr(Qureg qureg, std::vector<int> targets, DiagMatr matrix);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
 /// @see applyDiagMatr()
 void applyDiagMatr(Qureg qureg, std::vector<int> targets, DiagMatr matrix);
 
@@ -1013,22 +813,6 @@ void applyMultiControlledDiagMatr(Qureg qureg, std::vector<int> controls, std::v
 /// @cppvectoroverload
 /// @see applyMultiStateControlledDiagMatr()
 void applyMultiStateControlledDiagMatr(Qureg qureg, std::vector<int> controls, std::vector<int> states, std::vector<int> targets, DiagMatr matrix);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see multiplyDiagMatrPower()
-void multiplyDiagMatrPower(Qureg qureg, std::vector<int> targets, DiagMatr matrix, qcomp exponent);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see postMultiplyDiagMatrPower()
-void postMultiplyDiagMatrPower(Qureg qureg, std::vector<int> targets, DiagMatr matrix, qcomp exponent);
 
 
 /// @notyettested
@@ -1079,33 +863,6 @@ void applyMultiStateControlledDiagMatrPower(Qureg qureg, std::vector<int> contro
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @notyetvalidated
-/// @see
-/// - multiplyCompMatr1
-void multiplyFullStateDiagMatr(Qureg qureg, FullStateDiagMatr matrix);
-
-
-/// @notyetdoced
-/// @notyetvalidated
-/// @see
-/// - multiplyCompMatr1
-/// - applyDiagMatrPower
-void multiplyFullStateDiagMatrPower(Qureg qureg, FullStateDiagMatr matrix, qcomp exponent);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyFullStateDiagMatr(Qureg qureg, FullStateDiagMatr matrix);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyFullStateDiagMatrPower(Qureg qureg, FullStateDiagMatr matrix, qcomp exponent);
 
 
 /// @notyetdoced
@@ -1267,17 +1024,6 @@ extern "C" {
 #endif
 
 
-/// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplySwap(Qureg qureg, int qubit1, int qubit2);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplySwap(Qureg qureg, int qubit1, int qubit2);
-
-
 /** Applies a SWAP gate between @p qubit1 and @p qubit2 of @p qureg.
  * 
  * @diagram
@@ -1396,42 +1142,6 @@ void applyMultiStateControlledSqrtSwap(Qureg qureg, std::vector<int> controls, s
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see multiplyCompMatr1()
-void multiplyPauliX(Qureg qureg, int target);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see multiplyCompMatr1()
-void multiplyPauliY(Qureg qureg, int target);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see multiplyCompMatr1()
-void multiplyPauliZ(Qureg qureg, int target);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see postMultiplyCompMatr1()
-void postMultiplyPauliX(Qureg qureg, int target);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see postMultiplyCompMatr1()
-void postMultiplyPauliY(Qureg qureg, int target);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @see postMultiplyCompMatr1()
-void postMultiplyPauliZ(Qureg qureg, int target);
 
 
 /// @notyetdoced
@@ -1557,17 +1267,6 @@ void applyMultiStateControlledPauliZ(Qureg qureg, std::vector<int> controls, std
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplyPauliStr(Qureg qureg, PauliStr str);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyPauliStr(Qureg qureg, PauliStr str);
 
 
 /// @notyetdoced
@@ -1940,7 +1639,7 @@ void applyMultiStateControlledRotateAroundAxis(Qureg qureg, std::vector<int> ctr
 
 
 /** 
- * @defgroup op_pauligadget Pauli gadgets
+ * @defgroup op_pauligadget PauliStr gadgets
  * @brief Functions for applying many-qubit rotations around arbitrary PauliStr.
  * @{
  */
@@ -1949,19 +1648,6 @@ void applyMultiStateControlledRotateAroundAxis(Qureg qureg, std::vector<int> ctr
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @see 
-/// - multiplyCompMatr1()
-/// - applyPauliGadget()
-void multiplyPauliGadget(Qureg qureg, PauliStr str, qreal angle);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyPauliGadget(Qureg qureg, PauliStr str, qreal angle);
 
 
 /** @notyetdoced
@@ -2088,19 +1774,6 @@ void applyMultiStateControlledPauliGadget(Qureg qureg, std::vector<int> controls
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @see 
-/// - multiplyCompMatr1()
-/// - applyPhaseGadget
-void multiplyPhaseGadget(Qureg qureg, int* targets, int numTargets, qreal angle);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyPhaseGadget(Qureg qureg, int* targets, int numTargets, qreal angle);
 
 
 /** @notyetdoced
@@ -2371,22 +2044,6 @@ void applyMultiQubitPhaseShift(Qureg qureg, int* targets, int numTargets, qreal 
 /// @notyetvalidated
 /// @notyetdoced
 /// @cppvectoroverload
-/// @see multiplyPhaseGadget()
-void multiplyPhaseGadget(Qureg qureg, std::vector<int> targets, qreal angle);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see postMultiplyPhaseGadget()
-void postMultiplyPhaseGadget(Qureg qureg, std::vector<int> targets, qreal angle);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
 /// @see applyPhaseGadget()
 void applyPhaseGadget(Qureg qureg, std::vector<int> targets, qreal angle);
 
@@ -2438,8 +2095,8 @@ void applyMultiQubitPhaseShift(Qureg qureg, std::vector<int> targets, qreal angl
 
 
 /** 
- * @defgroup op_paulistrsum PauliStrSum
- * @brief Functions for applying, exponentiating or Trotterising a weigthed sum of Pauli tensors.
+ * @defgroup op_paulistrsum PauliStrSum gadgets
+ * @brief Functions for apply Trotterised exponentials of weighted sums of Pauli tensors.
  * @{
  */
 
@@ -2447,18 +2104,6 @@ void applyMultiQubitPhaseShift(Qureg qureg, std::vector<int> targets, qreal angl
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
-/// @notyetdoced
-/// @notyetvalidated
-/// @see multiplyCompMatr1()
-void multiplyPauliStrSum(Qureg qureg, PauliStrSum sum, Qureg workspace);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyPauliStrSum(Qureg qureg, PauliStrSum sum, Qureg workspace);
 
 
 /** @notyettested
@@ -2742,17 +2387,6 @@ extern "C" {
 
 
 /// @notyetdoced
-/// @see multiplyCompMatr1()
-void multiplyMultiQubitNot(Qureg qureg, int* targets, int numTargets);
-
-
-/// @notyetdoced
-/// @notyettested
-/// @notyetvalidated
-void postMultiplyMultiQubitNot(Qureg qureg, int* targets, int numTargets);
-
-
-/// @notyetdoced
 void applyMultiQubitNot(Qureg qureg, int* targets, int numTargets);
 
 
@@ -2776,22 +2410,6 @@ void applyMultiStateControlledMultiQubitNot(Qureg qureg, int* controls, int* sta
 #endif
 
 #ifdef __cplusplus
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see multiplyMultiQubitNot()
-void multiplyMultiQubitNot(Qureg qureg, std::vector<int> targets);
-
-
-/// @notyettested
-/// @notyetvalidated
-/// @notyetdoced
-/// @cppvectoroverload
-/// @see postMultiplyMultiQubitNot()
-void postMultiplyMultiQubitNot(Qureg qureg, std::vector<int> targets);
 
 
 /// @notyettested
