+++ quest/src/api/CMakeLists.txt
@@ -8,6 +8,7 @@ target_sources(QuEST
   initialisations.cpp
   matrices.cpp
   modes.cpp
+  multiplication.cpp
   operations.cpp
   paulis.cpp
   qureg.cpp
