+++ quest/src/api/operations.cpp
@@ -8,6 +8,7 @@
  */
 
 #include "quest/include/qureg.h"
+#include "quest/include/paulis.h"
 #include "quest/include/matrices.h"
 #include "quest/include/operations.h"
 #include "quest/include/calculations.h"
@@ -75,29 +76,6 @@ void validateAndApplyAnyCtrlAnyTargUnitaryMatrix(Qureg qureg, int* ctrls, int* s
 
 extern "C" {
 
-void multiplyCompMatr1(Qureg qureg, int target, CompMatr1 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_target(qureg, target, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-
-    bool conj = false;
-    bool transp = false;
-    localiser_statevec_anyCtrlOneTargDenseMatr(qureg, {}, {}, target, matrix, conj, transp);
-}
-
-void postMultiplyCompMatr1(Qureg qureg, int target, CompMatr1 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_target(qureg, target, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-    
-    // rho matrix ~ transpose(rho) (x) I ||rho>>
-    bool conj = false;
-    bool transp = true;
-    int qubit = util_getBraQubit(target, qureg);
-    localiser_statevec_anyCtrlOneTargDenseMatr(qureg, {}, {}, qubit, matrix, conj, transp);
-}
-
 void applyCompMatr1(Qureg qureg, int target, CompMatr1 matrix) {
 
     validateAndApplyAnyCtrlAnyTargUnitaryMatrix(qureg, nullptr, nullptr, 0, &target, 1, matrix, __func__);
@@ -138,32 +116,6 @@ void applyMultiStateControlledCompMatr1(Qureg qureg, vector<int> controls, vecto
  */
 
 extern "C" {
-
-void multiplyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_twoTargets(qureg, target1, target2, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-    validate_mixedAmpsFitInNode(qureg, 2, __func__);
-
-    bool conj = false;
-    bool transp = false;
-    localiser_statevec_anyCtrlTwoTargDenseMatr(qureg, {}, {}, target1, target2, matrix, conj, transp);
-}
-
-void postMultiplyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_twoTargets(qureg, target1, target2, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-    validate_mixedAmpsFitInNode(qureg, 2, __func__);
-
-    // rho matrix ~ transpose(rho) (x) I ||rho>>
-    bool conj = false;
-    bool transp = true;
-    int qubit1 = util_getBraQubit(target1, qureg);
-    int qubit2 = util_getBraQubit(target2, qureg);
-    localiser_statevec_anyCtrlTwoTargDenseMatr(qureg, {}, {}, qubit1, qubit2, matrix, conj, transp);
-}
 
 void applyCompMatr2(Qureg qureg, int target1, int target2, CompMatr2 matrix) {
 
@@ -210,31 +162,6 @@ void applyMultiStateControlledCompMatr2(Qureg qureg, vector<int> controls, vecto
 
 extern "C" {
 
-void multiplyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-    validate_mixedAmpsFitInNode(qureg, numTargets, __func__);
-
-    bool conj = false;
-    bool transp = false;
-    localiser_statevec_anyCtrlAnyTargDenseMatr(qureg, {}, {}, util_getVector(targets, numTargets), matrix, conj, transp);
-}
-
-void postMultiplyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-    validate_mixedAmpsFitInNode(qureg, numTargets, __func__);
-
-    // rho matrix ~ transpose(rho) (x) I ||rho>>
-    bool conj = false;
-    bool transp = true;
-    auto qubits = util_getBraQubits(util_getVector(targets, numTargets), qureg);
-    localiser_statevec_anyCtrlAnyTargDenseMatr(qureg, {}, {}, qubits, matrix, conj, transp);
-}
-
 void applyCompMatr(Qureg qureg, int* targets, int numTargets, CompMatr matrix) {
 
     validateAndApplyAnyCtrlAnyTargUnitaryMatrix(qureg, nullptr, nullptr, 0, targets, numTargets, matrix, __func__);
@@ -256,16 +183,6 @@ void applyMultiStateControlledCompMatr(Qureg qureg, int* controls, int* states, 
 }
 
 } // end de-mangler
-
-void multiplyCompMatr(Qureg qureg, vector<int> targets, CompMatr matr) {
-
-    multiplyCompMatr(qureg, targets.data(), targets.size(), matr);
-}
-
-void postMultiplyCompMatr(Qureg qureg, vector<int> targets, CompMatr matr) {
-
-    postMultiplyCompMatr(qureg, targets.data(), targets.size(), matr);
-}
 
 void applyCompMatr(Qureg qureg, vector<int> targets, CompMatr matr) {
 
@@ -295,26 +212,6 @@ void applyMultiStateControlledCompMatr(Qureg qureg, vector<int> controls, vector
  */
 
 extern "C" {
-
-void multiplyDiagMatr1(Qureg qureg, int target, DiagMatr1 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_target(qureg, target, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-
-    bool conj = false;
-    localiser_statevec_anyCtrlOneTargDiagMatr(qureg, {}, {}, target, matrix, conj);
-}
-
-void postMultiplyDiagMatr1(Qureg qureg, int target, DiagMatr1 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_target(qureg, target, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-
-    bool conj = false;
-    int qubit = util_getBraQubit(target, qureg);
-    localiser_statevec_anyCtrlOneTargDiagMatr(qureg, {}, {}, qubit, matrix, conj);
-}
 
 void applyDiagMatr1(Qureg qureg, int target, DiagMatr1 matrix) {
 
@@ -356,27 +253,6 @@ void applyMultiStateControlledDiagMatr1(Qureg qureg, vector<int> controls, vecto
  */
 
 extern "C" {
-
-void multiplyDiagMatr2(Qureg qureg, int target1, int target2, DiagMatr2 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_twoTargets(qureg, target1, target2, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-
-    bool conj = false;
-    localiser_statevec_anyCtrlTwoTargDiagMatr(qureg, {}, {}, target1, target2, matrix, conj);
-}
-
-void postMultiplyDiagMatr2(Qureg qureg, int target1, int target2, DiagMatr2 matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_twoTargets(qureg, target1, target2, __func__);
-    validate_matrixFields(matrix, __func__); // matrix can be non-unitary
-
-    bool conj = false;
-    int qubit1 = util_getBraQubit(target1, qureg);
-    int qubit2 = util_getBraQubit(target2, qureg);
-    localiser_statevec_anyCtrlTwoTargDiagMatr(qureg, {}, {}, qubit1, qubit2, matrix, conj);
-}
 
 void applyDiagMatr2(Qureg qureg, int target1, int target2, DiagMatr2 matrix) {
 
@@ -423,29 +299,6 @@ void applyMultiStateControlledDiagMatr2(Qureg qureg, vector<int> controls, vecto
 
 extern "C" {
 
-void multiplyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-
-    bool conj = false;
-    qcomp exponent = 1;
-    auto qubits = util_getVector(targets, numTargets);
-    localiser_statevec_anyCtrlAnyTargDiagMatr(qureg, {}, {}, qubits, matrix, exponent, conj);
-}
-
-void postMultiplyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-
-    bool conj = false;
-    qcomp exponent = 1;
-    auto qubits = util_getBraQubits(util_getVector(targets, numTargets), qureg);
-    localiser_statevec_anyCtrlAnyTargDiagMatr(qureg, {}, {}, qubits, matrix, exponent, conj);
-}
-
 void applyDiagMatr(Qureg qureg, int* targets, int numTargets, DiagMatr matrix) {
 
     validateAndApplyAnyCtrlAnyTargUnitaryMatrix(qureg, nullptr, nullptr, 0, targets, numTargets, matrix, __func__);
@@ -467,16 +320,6 @@ void applyMultiStateControlledDiagMatr(Qureg qureg, int* controls, int* states, 
 }
 
 } // end de-mangler
-
-void multiplyDiagMatr(Qureg qureg, vector<int> targets, DiagMatr matrix) {
-
-    multiplyDiagMatr(qureg, targets.data(), targets.size(), matrix);
-}
-
-void postMultiplyDiagMatr(Qureg qureg, vector<int> targets, DiagMatr matrix) {
-
-    postMultiplyDiagMatr(qureg, targets.data(), targets.size(), matrix);
-}
 
 void applyDiagMatr(Qureg qureg, vector<int> targets, DiagMatr matrix) {
 
@@ -504,34 +347,11 @@ void applyMultiStateControlledDiagMatr(Qureg qureg, vector<int> controls, vector
 /*
  * DiagMatrPower
  *
- * which still (except for multiply) assert unitarity,
- * even though a non-real exponent is possible
+ * which still assert unitarity despite that passing
+ * a non-real exponent is permitted
  */
 
 extern "C" {
-
-void multiplyDiagMatrPower(Qureg qureg, int* targets, int numTargets, DiagMatr matrix, qcomp exponent) {
-    validate_quregFields(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-    validate_matrixExpIsNonDiverging(matrix, exponent, __func__); // harmlessly re-validates fields and is-sync
-
-    bool conj = false;
-    auto qubits = util_getVector(targets, numTargets);
-    localiser_statevec_anyCtrlAnyTargDiagMatr(qureg, {}, {}, qubits, matrix, exponent, conj);
-}
-
-void postMultiplyDiagMatrPower(Qureg qureg, int* targets, int numTargets, DiagMatr matrix, qcomp exponent) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-    validate_matrixDimMatchesTargets(matrix, numTargets, __func__); // also validates fields and is-sync, but not unitarity
-    validate_matrixExpIsNonDiverging(matrix, exponent, __func__); // harmlessly re-validates fields and is-sync
-
-    bool conj = false;
-    auto qubits = util_getBraQubits(util_getVector(targets, numTargets), qureg);
-    localiser_statevec_anyCtrlAnyTargDiagMatr(qureg, {}, {}, qubits, matrix, exponent, conj);
-}
 
 void applyDiagMatrPower(Qureg qureg, int* targets, int numTargets, DiagMatr matrix, qcomp exponent)  {
     validate_quregFields(qureg, __func__);
@@ -610,16 +430,6 @@ void applyMultiStateControlledDiagMatrPower(Qureg qureg, int* controls, int* sta
 
 } // end de-mangler
 
-void multiplyDiagMatrPower(Qureg qureg, vector<int> targets, DiagMatr matrix, qcomp exponent) {
-
-    multiplyDiagMatrPower(qureg, targets.data(), targets.size(), matrix, exponent);
-}
-
-void postMultiplyDiagMatrPower(Qureg qureg, vector<int> targets, DiagMatr matrix, qcomp exponent) {
-
-    postMultiplyDiagMatrPower(qureg, targets.data(), targets.size(), matrix, exponent);
-}
-
 void applyDiagMatrPower(Qureg qureg, vector<int> targets, DiagMatr matrix, qcomp exponent) {
 
     applyDiagMatrPower(qureg, targets.data(), targets.size(), matrix, exponent);
@@ -648,53 +458,6 @@ void applyMultiStateControlledDiagMatrPower(Qureg qureg, vector<int> controls, v
  */
 
 extern "C" {
-
-void multiplyFullStateDiagMatr(Qureg qureg, FullStateDiagMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_matrixFields(matrix, __func__);
-    validate_matrixAndQuregAreCompatible(matrix, qureg, false, __func__); // matrix can be non-unitary
-
-    multiplyFullStateDiagMatrPower(qureg, matrix, 1); // harmlessly re-validates
-}
-
-void multiplyFullStateDiagMatrPower(Qureg qureg, FullStateDiagMatr matrix, qcomp exponent) {
-    validate_quregFields(qureg, __func__);
-    validate_matrixFields(matrix, __func__);
-    validate_matrixAndQuregAreCompatible(matrix, qureg, false, __func__); // matrix can be non-unitary
-    validate_matrixExpIsNonDiverging(matrix, exponent, __func__);
-
-    // rho -> matrix^exponent rho
-    bool leftMultiply = true;
-    bool rightMultiply = false;
-    bool rightConj = false;
-
-    (qureg.isDensityMatrix)?
-        localiser_densmatr_allTargDiagMatr(qureg, matrix, exponent, leftMultiply, rightMultiply, rightConj):
-        localiser_statevec_allTargDiagMatr(qureg, matrix, exponent);
-}
-
-void postMultiplyFullStateDiagMatr(Qureg qureg, FullStateDiagMatr matrix) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_matrixFields(matrix, __func__);
-    validate_matrixAndQuregAreCompatible(matrix, qureg, false, __func__); // matrix can be non-unitary
-
-    postMultiplyFullStateDiagMatrPower(qureg, matrix, 1); // harmlessly re-validates
-}
-
-void postMultiplyFullStateDiagMatrPower(Qureg qureg, FullStateDiagMatr matrix, qcomp exponent) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_matrixFields(matrix, __func__);
-    validate_matrixAndQuregAreCompatible(matrix, qureg, false, __func__); // matrix can be non-unitary
-    validate_matrixExpIsNonDiverging(matrix, exponent, __func__);
-
-    // rho -> rho matrix^exponent
-    bool leftMultiply = false;
-    bool rightMultiply = true;
-    bool rightConj = false;
-    localiser_densmatr_allTargDiagMatr(qureg, matrix, exponent, leftMultiply, rightMultiply, rightConj);
-}
 
 void applyFullStateDiagMatr(Qureg qureg, FullStateDiagMatr matrix) {
     validate_quregFields(qureg, __func__);
@@ -890,23 +653,6 @@ void applyMultiStateControlledHadamard(Qureg qureg, vector<int> controls, vector
 
 extern "C" {
 
-void multiplySwap(Qureg qureg, int qubit1, int qubit2) {
-    validate_quregFields(qureg, __func__);
-    validate_twoTargets(qureg, qubit1, qubit2, __func__);
-
-    localiser_statevec_anyCtrlSwap(qureg, {}, {}, qubit1, qubit2);
-}
-
-void postMultiplySwap(Qureg qureg, int qubit1, int qubit2) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_twoTargets(qureg, qubit1, qubit2, __func__);
-
-    qubit1 = util_getBraQubit(qubit1, qureg);
-    qubit2 = util_getBraQubit(qubit2, qureg);
-    localiser_statevec_anyCtrlSwap(qureg, {}, {}, qubit1, qubit2);
-}
-
 void applySwap(Qureg qureg, int qubit1, int qubit2) {
     validate_quregFields(qureg, __func__);
     validate_twoTargets(qureg, qubit1, qubit2, __func__);
@@ -1039,61 +785,6 @@ void applyMultiStateControlledSqrtSwap(Qureg qureg, vector<int> controls, vector
  */
 
 extern "C" {
-
-void multiplyPauliX(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    PauliStr str = getPauliStr("X", {target});
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
-
-void multiplyPauliY(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    PauliStr str = getPauliStr("Y", {target});
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
-
-void multiplyPauliZ(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    PauliStr str = getPauliStr("Z", {target});
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
-
-void postMultiplyPauliX(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    PauliStr str = getPauliStr("X", {target});
-    str = paulis_getShiftedPauliStr(str, qureg.numQubits);
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
-
-void postMultiplyPauliY(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    qcomp factor = -1; // undo transpose
-    PauliStr str = getPauliStr("Y", {target});
-    str = paulis_getShiftedPauliStr(str, qureg.numQubits);
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str, factor);
-}
-
-void postMultiplyPauliZ(Qureg qureg, int target) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_target(qureg, target, __func__);
-
-    PauliStr str = getPauliStr("Z", {target});
-    str = paulis_getShiftedPauliStr(str, qureg.numQubits);
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
 
 void applyPauliX(Qureg qureg, int target) {
     validate_quregFields(qureg, __func__);
@@ -1238,23 +929,6 @@ void applyMultiStateControlledPauliZ(Qureg qureg, vector<int> controls, vector<i
 
 extern "C" {
 
-void multiplyPauliStr(Qureg qureg, PauliStr str) {
-    validate_quregFields(qureg, __func__);
-    validate_pauliStrTargets(qureg, str, __func__);
-
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str);
-}
-
-void postMultiplyPauliStr(Qureg qureg, PauliStr str) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_pauliStrTargets(qureg, str, __func__);
-
-    qcomp factor = paulis_hasOddNumY(str)? -1 : 1; // undo transpose
-    str = paulis_getShiftedPauliStr(str, qureg.numQubits);
-    localiser_statevec_anyCtrlPauliTensor(qureg, {}, {}, str, factor);
-}
-
 void applyPauliStr(Qureg qureg, PauliStr str) {
     validate_quregFields(qureg, __func__);
     validate_pauliStrTargets(qureg, str, __func__);
@@ -1329,52 +1003,6 @@ void applyMultiStateControlledPauliStr(Qureg qureg, vector<int> controls, vector
  */
 
 extern "C" {
-
-void multiplyPauliStrSum(Qureg qureg, PauliStrSum sum, Qureg workspace) {
-    validate_quregFields(qureg, __func__);
-    validate_quregFields(workspace, __func__);
-    validate_quregCanBeWorkspace(qureg, workspace, __func__);
-    validate_pauliStrSumFields(sum, __func__);
-    validate_pauliStrSumTargets(sum, qureg, __func__);
-
-    // clone qureg to workspace, set qureg to blank
-    localiser_statevec_setQuregToSuperposition(0, workspace, 1, qureg, 0, qureg);
-    localiser_statevec_initUniformState(qureg, 0);
-
-    // left-multiply each term in-turn, mixing into output qureg, then undo using idempotency
-    for (qindex i=0; i<sum.numTerms; i++) {
-        localiser_statevec_anyCtrlPauliTensor(workspace, {}, {}, sum.strings[i]);
-        localiser_statevec_setQuregToSuperposition(1, qureg, sum.coeffs[i], workspace, 0, workspace);
-        localiser_statevec_anyCtrlPauliTensor(workspace, {}, {}, sum.strings[i]);
-    }
-
-    // workspace -> qureg, and qureg -> sum * qureg
-}
-
-void postMultiplyPauliStrSum(Qureg qureg, PauliStrSum sum, Qureg workspace) {
-    validate_quregFields(qureg, __func__);
-    validate_quregFields(workspace, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_quregCanBeWorkspace(qureg, workspace, __func__);
-    validate_pauliStrSumFields(sum, __func__);
-    validate_pauliStrSumTargets(sum, qureg, __func__);
-
-    // clone qureg to workspace, set qureg to blank
-    localiser_statevec_setQuregToSuperposition(0, workspace, 1, qureg, 0, qureg);
-    localiser_statevec_initUniformState(qureg, 0);
-
-    // post-multiply each term in-turn, mixing into output qureg, then undo using idempotency
-    for (qindex i=0; i<sum.numTerms; i++) {
-        PauliStr str =  paulis_getShiftedPauliStr(sum.strings[i], qureg.numQubits);
-        qcomp factor = paulis_hasOddNumY(str)? -1 : 1; // undoes transpose
-
-        localiser_statevec_anyCtrlPauliTensor(workspace, {}, {}, str, factor);
-        localiser_statevec_setQuregToSuperposition(1, qureg, sum.coeffs[i], workspace, 0, workspace);
-        localiser_statevec_anyCtrlPauliTensor(workspace, {}, {}, str, factor);
-    }
-
-    // workspace -> qureg, and qureg -> sum * qureg
-}
 
 void internal_applyFirstOrderTrotterRepetition(
     Qureg qureg, vector<int>& ketCtrls, vector<int>& braCtrls,
@@ -1522,8 +1150,6 @@ void applyMultiStateControlledTrotterizedPauliStrSumGadget(Qureg qureg, vector<i
  */
 
 extern "C" {
-
-// don't think users will ever want to left-multiply only
 
 void applyRotateX(Qureg qureg, int target, qreal angle) {
     validate_quregFields(qureg, __func__);
@@ -1734,25 +1360,6 @@ void applyMultiStateControlledRotateAroundAxis(Qureg qureg, vector<int> ctrls, v
 
 extern "C" {
 
-void multiplyPauliGadget(Qureg qureg, PauliStr str, qreal angle) {
-    validate_quregFields(qureg, __func__);
-    validate_pauliStrTargets(qureg, str, __func__);
-
-    qreal phase = util_getPhaseFromGateAngle(angle);
-    localiser_statevec_anyCtrlPauliGadget(qureg, {}, {}, str, phase);
-}
-
-void postMultiplyPauliGadget(Qureg qureg, PauliStr str, qreal angle) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_pauliStrTargets(qureg, str, __func__);
-
-    qreal factor = paulis_hasOddNumY(str)? -1 : 1;
-    qreal phase = factor * util_getPhaseFromGateAngle(angle);
-    str = paulis_getShiftedPauliStr(str, qureg.numQubits);
-    localiser_statevec_anyCtrlPauliGadget(qureg, {}, {}, str, phase);
-}
-
 void applyPauliGadget(Qureg qureg, PauliStr str, qreal angle) {
     validate_quregFields(qureg, __func__);
     validate_pauliStrTargets(qureg, str, __func__);
@@ -1841,25 +1448,6 @@ void applyMultiStateControlledPauliGadget(Qureg qureg, vector<int> controls, vec
 
 extern "C" {
 
-void multiplyPhaseGadget(Qureg qureg, int* targets, int numTargets, qreal angle) {
-    validate_quregFields(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-
-    qreal phase = util_getPhaseFromGateAngle(angle);
-    auto qubits = util_getVector(targets, numTargets);
-    localiser_statevec_anyCtrlPhaseGadget(qureg, {}, {}, qubits, phase);
-}
-
-void postMultiplyPhaseGadget(Qureg qureg, int* targets, int numTargets, qreal angle) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-
-    qreal phase = util_getPhaseFromGateAngle(angle);
-    auto qubits = util_getBraQubits(util_getVector(targets, numTargets), qureg);
-    localiser_statevec_anyCtrlPhaseGadget(qureg, {}, {}, qubits, phase);
-}
-
 void applyPhaseGadget(Qureg qureg, int* targets, int numTargets, qreal angle) {
     validate_quregFields(qureg, __func__);
     validate_targets(qureg, targets, numTargets, __func__);
@@ -1905,16 +1493,6 @@ void applyMultiStateControlledPhaseGadget(Qureg qureg, int* controls, int* state
 }
 
 } // end de-mangler
-
-void multiplyPhaseGadget(Qureg qureg, vector<int> targets, qreal angle) {
-
-    multiplyPhaseGadget(qureg, targets.data(), targets.size(), angle);
-}
-
-void postMultiplyPhaseGadget(Qureg qureg, vector<int> targets, qreal angle) {
-
-    postMultiplyPhaseGadget(qureg, targets.data(), targets.size(), angle);
-}
 
 void applyPhaseGadget(Qureg qureg, vector<int> targets, qreal angle) {
 
@@ -2026,29 +1604,10 @@ void applyMultiQubitPhaseFlip(Qureg qureg, vector<int> targets) {
 
 
 /*
- * many-qubit CNOTs
+ * many-qubit NOTs and CNOTs
  */
 
 extern "C" {
-
-void multiplyMultiQubitNot(Qureg qureg, int* targets, int numTargets) {
-    validate_quregFields(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-
-    // harmlessly re-validates
-    PauliStr str = getPauliStr(std::string(numTargets, 'X'), targets, numTargets);
-    multiplyPauliStr(qureg, str);
-}
-
-void postMultiplyMultiQubitNot(Qureg qureg, int* targets, int numTargets) {
-    validate_quregFields(qureg, __func__);
-    validate_quregIsDensityMatrix(qureg, __func__);
-    validate_targets(qureg, targets, numTargets, __func__);
-
-    // harmlessly re-validates
-    PauliStr str = getPauliStr(std::string(numTargets, 'X'), targets, numTargets);
-    postMultiplyPauliStr(qureg, str);
-}
 
 void applyMultiQubitNot(Qureg qureg, int* targets, int numTargets) {
     validate_quregFields(qureg, __func__);
@@ -2087,16 +1646,6 @@ void applyMultiStateControlledMultiQubitNot(Qureg qureg, int* controls, int* sta
 }
 
 } // end de-mangler
-
-void multiplyMultiQubitNot(Qureg qureg, vector<int> targets) {
-
-    multiplyMultiQubitNot(qureg, targets.data(), targets.size());
-}
-
-void postMultiplyMultiQubitNot(Qureg qureg, vector<int> targets) {
-
-    postMultiplyMultiQubitNot(qureg, targets.data(), targets.size());
-}
 
 void applyMultiQubitNot(Qureg qureg, vector<int> targets) {
