+++ tests/unit/multiplication.cpp
@@ -1,0 +1,7 @@
+/*
+ * Testing of the multiplication API module is actually 
+ * performed in operations.cpp, in addition to testing
+ * the operations module, since their testing logics are
+ * inextricable. This file exists only to redirect the
+ * confused reader searching for the multiplication tests.
+ */
